Prüfung ob LOG_FILE existiert

+++ bin/yavdr_funcs.sh
@@ -27,7 +27,7 @@ f_logger() {
       -o|--osd)
         parm='--stderr' ; shift
         #/usr/bin/vdr-dbus-send /Skin skin.QueueMessage string:"$*" ;;
-        f_svdrpsend MESG "$*"
+        f_svdrpsend MESG "$*" ;;
     esac
     logger "$parm" --tag 'yaVDR' "[$$] ${SELF_NAME}: $*"
   fi
@@ -36,19 +36,19 @@ f_logger() {
 f_log() {  # Gibt die Meldung auf der Konsole und im Syslog aus
   [[ -t 1 ]] && echo "$*"      # Konsole
   logger -t "$SELF_NAME" "$*"  # Syslog
-  [[ -w "${LOG_FILE:-/dev/null}" ]] && echo "$*" >> "$LOG_FILE"  # Log in Datei
+  [[ -n "$LOG_FILE" && -w "$LOG_FILE" ]] && echo "$*" >> "$LOG_FILE"  # Log in Datei
 }
 
 f_log2() {                                           
   local data=("${@:-$(</dev/stdin)}")               # Akzeptiert Parameter und via stdin (|)
   [[ -t 1 ]] && printf '%s\n' "${data[@]}"          # Konsole falls verbunden
   logger -t "$SELF_NAME" "${data[@]}"               # Systemlog
-  [[ -w "${LOG_FILE:-/dev/null}" ]] && printf '%s\n' "${data[@]}" >> "$LOG_FILE"  # Log-Datei
+  [[ -n "$LOG_FILE" && -w "$LOG_FILE" ]] && printf '%s\n' "${data[@]}" >> "$LOG_FILE"  # Log-Datei
 }
 
 f_rotate_log() {  # Log rotieren wenn zu groß
   local file="${LOG_FILE:-$1}"
-  if [[ -w "$file" ]] ; then  # Datei Existiert und hat Schreibrechte
+  if [[ -n "$file" && -w "$file" ]] ; then  # Datei Existiert und hat Schreibrechte
     FILE_SIZE="$(stat -c %s "$file" 2>/dev/null)"
     [[ ${FILE_SIZE:-51201} -gt ${MAX_LOG_SIZE:-51200} ]] && mv --force "$file" "${file}.old"
   fi
